chore(popup): move link import
