+++ packages/popup/src/components/FileInput.tsx
@@ -3,12 +3,12 @@ import type { JSXInternal } from "preact/src/jsx";
 
 import { POPUP_ROUTES } from "@worm/shared";
 
+import { useLanguage } from "../lib/language";
 import {
   CAN_UPLOAD_PARAMETER,
   NOTIFY_PARAMETER,
   canUploadDirect,
 } from "../lib/routes";
-import { useLanguage } from "../lib/language";
 
 type FileUploadProps = {
   onChange: (event: JSXInternal.TargetedInputEvent<HTMLInputElement>) => void;
